Dictionary integration
